Consolidate codicon files to vs/base/parts/codicon
