+++ src/vs/base/browser/ui/highlightedlabel/highlightedLabel.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as objects from 'vs/base/common/objects';
-import { renderCodicons } from 'vs/base/common/codicons';
+import { renderCodicons } from 'vs/base/parts/codicon/common/codicons';
 import { escape } from 'vs/base/common/strings';
 
 export interface IHighlight {
