+++ src/vs/base/parts/codicon/browser/codicon-animations.css
