+++ src/vs/base/parts/codicon/browser/codicon.css
