+++ src/vs/base/parts/codicon/browser/codicon.ttf
