+++ src/vs/base/parts/codicon/common/codicon.ts
