+++ src/vs/base/parts/codicon/common/codicons.ts
