+++ src/vs/workbench/api/common/extHostTypes.ts
@@ -14,7 +14,7 @@ import { generateUuid } from 'vs/base/common/uuid';
 import * as vscode from 'vscode';
 import { FileSystemProviderErrorCode, markAsFileSystemProviderError } from 'vs/platform/files/common/files';
 import { RemoteAuthorityResolverErrorCode } from 'vs/platform/remote/common/remoteAuthorityResolver';
-import { escapeCodicons } from 'vs/base/common/codicons';
+import { escapeCodicons } from 'vs/base/parts/codicon/common/codicons';
 
 function es5ClassCompat(target: Function): any {
 	///@ts-ignore
